Move options tests to separate dir
