+++ test/options/alias.js
