+++ test/options/keywordCase.js
