+++ test/options/keywordPosition.js
