+++ test/options/newline.js
