+++ test/options/newlineBeforeParen.js
