+++ test/options/newlineBeforeSemicolon.js
